+++ src/components/Home/Authenticated/index.ts
@@ -1,0 +1,4 @@
+export { AuthenticatedContainer } from "./AuthenticatedContainer";
+export { TripNav } from "./TripNav";
+export { NewTripButton } from "./NewTripButton";
+export { TripsContainer } from "./TripsContainer";
