+++ src/components/Home/HomeContainer.tsx
@@ -1,5 +1,5 @@
-import { UnauthenticatedContainer } from "./Unauthenticated/UnauthenticatedContainer";
-import { AuthenticatedContainer } from "./Authenticated/AuthenticatedContainer";
+import { UnauthenticatedContainer } from "./Unauthenticated";
+import { AuthenticatedContainer } from "./Authenticated";
 import { useSession } from "next-auth/react";
 
 export const HomeContainer = () => {
